LayoutTests/http/tests/canvas/canvas-tainted-after-draw-image-expected.txt

CONSOLE MESSAGE: Unable to get image data from canvas because the canvas has been tainted by cross-origin data.
Tainting works correctly.
Tainting works correctly.
Tainting works correctly.
Tainting works correctly.
PASS successfullyParsed is true

TEST COMPLETE
